+++ Changelog.md
@@ -22,6 +22,7 @@ Breaking changes:
 
 
  * Language Feature: When overriding a function or modifier, the new keyword ``override`` must be used. When overriding a function or modifier defined in multiple parallel bases, all bases must be listed in parentheses after the keyword like so: ``override(Base1, Base2)``
+ * Type Checker: Disallows nested mappings that are requested to live outside storage.
 
 Language Features:
  * Allow global enums and structs.